Extract relay connection to module
