+++ lib/graphql/types/relay.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
+require "graphql/types/relay/connection_behaviors"
 require "graphql/types/relay/base_field"
 require "graphql/types/relay/base_object"
 require "graphql/types/relay/base_interface"
